ai-engine/requirements.txt
fastapi
uvicorn
langchain
transformers
pymongo
python-dotenv
requests
google-generativeai
